requirements.txt
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.95.0
h11==0.14.0
idna==3.4
passlib==1.7.4
pycparser==2.21
pydantic==1.10.7
PyJWT==2.6.0
pymongo==4.3.3
python-multipart==0.0.6
requests==2.28.2
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
psycopg2
pydantic[email]
pydantic[dotenv]
